- Limited amount of shards that can spawn on the ground - Reduced cooldown on weapons - Added chance of appearance to weapons and items - Added sound effects to buttons, enemy damage, exp pickup - Added background music in menu and ambient sound in gameplay - Fixed a banch of weapon/item upgrades missmatches - Rarity now applies changes on level up cards - Fixed unobtainable healing achievement - Changed armor rarity change from int to float

+++ Assets/Objects/Players/Scripts/PlayerStatsComponent.cs
@@ -225,7 +225,7 @@ namespace Objects.Players.Scripts
 			return _playerStats?.MovementSpeed ?? 0;
 		}
 
-		public int GetArmor()
+		public float GetArmor()
 		{
 			return _playerStats?.Armor ?? 0;
 		}
